Rename app to BCWallet
