+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.wallet",
+    package = "ca.bc.gov.bcwallet",
 )
 
 android_resource(
     name = "res",
-    package = "com.wallet",
+    package = "ca.bc.gov.bcwallet",
     res = "src/main/res",
 )
