+++ android/app/src/main/java/ca/bc/gov/bcwallet/MainActivity.java
@@ -1,4 +1,4 @@
-package com.wallet;
+package ca.bc.gov.bcBCWallet;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "wallet";
+    return "BCWallet";
   }
 }
