+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "wallet",
-  "displayName": "wallet"
+  "name": "BCWallet",
+  "displayName": "BCWallet"
 }
