+++ ios/BCWallet.xcodeproj/xcshareddata/xcschemes/wallet.xcscheme
