+++ ios/BCWallet.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
