+++ ios/BCWallet/AppDelegate.h
