+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/Contents.json
