+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/Icon-20@3x.png
