+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/Icon-29@2x.png
