+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/Icon-29@3x.png
