+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/Icon-40@3x.png
