+++ ios/BCWallet/Images.xcassets/AppIcon.appiconset/iTunesArtwork.png
