+++ ios/BCWallet/Images.xcassets/Contents.json
