+++ ios/BCWallet/Info.plist
