+++ ios/BCWallet/LaunchScreen.storyboard
