+++ ios/BCWalletTests/Info.plist
