+++ ios/BCWalletTests/walletTests.m
